Create basic etch-a-sketch functionality
Iterated through the array of grid items to add an event listener to each item. This event listener turns the item black when moused over by the user. This creates an etch-a-sketch like effect on the grid. The user does not yet have the ability to generate a grid of their chosen size. There is also an event listener and function that have been commented out that, if enabled, caused the squares to return to their original state after the mouse leaves the square.