Remove previous grid upon user input
The generateGrid function now removes the previous grid (if any) before generating the new grid from the user's input, so that the new grid items do not simply get added to the end of the old grid items. The grid's overall size and the number of columns are still hard-set, so there will still be 16 columns no matter what number the user inputs.