Enable user input of grid size
The button at the top of the page now produces a prompt that enables the user to specify how large they want each side of the square grid to be. The page still starts with a default 16x16 grid, then generates a new grid of the size asked for by the user. There are not yet any restrictions on the input type in the prompt.